reduce network size

+++ pax/ppo/ppo_gru.py
@@ -369,7 +369,7 @@ class PPO:
 
             # We pass through initial_hidden_state so its easy to batch memory
             key, subkey = jax.random.split(key)
-            dummy_obs = jnp.zeros(shape=obs_spec, dtype=jnp.float16)
+            dummy_obs = jnp.zeros(shape=obs_spec)
             dummy_obs = utils.add_batch_dim(dummy_obs)
             initial_params = network.init(
                 subkey, dummy_obs, initial_hidden_state
